Add files via upload
The idea for this small project came from the fact that there was not a proper way to reach out to people who for example like to play football. A couple of days after I made this, I saw on facebook that the 'Uletti' app came out and that was basically it for this project :).